Renamed managers to repositories for clarity
