+++ src/core/employee.h
@@ -4,11 +4,11 @@
 #include "order.h"
 #include "material.h"
 #include "reports.h"
-#include "managers/ordermanager.h"
-#include "managers/reportmanager.h"
-#include "managers/materialmanager.h"
+#include "repos/orderrepository.h"
+#include "repos/reportrepository.h"
+#include "repos/materialrepository.h"
 #include "utilities/clocks.h"
-#include "managers/employeemanager.h"
+#include "repos/employeerepository.h"
 
 #include <string>
 #include <map>
