+++ src/main.cpp
@@ -1,12 +1,12 @@
 #include "core/material.h"
 #include "core/client.h"
-#include "managers/clientmanager.h"
+#include "repos/clientrepository.h"
 #include "core/employee.h"
 #include "employees/administrator.h"
 #include "employees/receptionist.h"
 #include "employees/photographer.h"
 #include "core/order.h"
-#include "managers/employeemanager.h"
+#include "repos/employeerepository.h"
 #include "ui/uimanager.h"
 #include "utilities/clocks.h"
