+++ src/ui/uimanager.h
@@ -1,12 +1,12 @@
 #ifndef UIMANAGER_H
 #define UIMANAGER_H
 #include "ui.h"
-#include "managers/ordermanager.h"
-#include "managers/clientmanager.h"
-#include "managers/employeemanager.h"
-#include "managers/ordermanager.h"
-#include "managers/reportmanager.h"
-#include "managers/materialmanager.h"
+#include "repos/orderrepository.h"
+#include "repos/clientrepository.h"
+#include "repos/employeerepository.h"
+#include "repos/orderrepository.h"
+#include "repos/reportrepository.h"
+#include "repos/materialrepository.h"
 #include "employees/photographer.h"
 #include "employees/receptionist.h"
 #include "employees/administrator.h"
